Finally *hopefully* Fixed Move Code

+++ Team22258/src/main/java/org/firstinspires/ftc/team22258/teleop/opmode_TeleOp.java
@@ -29,10 +29,10 @@ public class opmode_TeleOp extends LinearOpMode {
   
   int ArmTiltPower;
   
-  int FLMotorPower;
-  int FRMotorPower;
-  int BLMotorPower;
-  int BRMotorPower;
+  double FLMotorPower;
+  double FRMotorPower;
+  double BLMotorPower;
+  double BRMotorPower;
   
   double MotorPowerNormalizer;
